refactor(atomix): nest domain specific classes
Nests the small `ValidationResult` and `NoopEntryValidator` classes
directly under `EntryValidator`, as they are types meant to be used only
in the context of the `EntryValidator` and no where else.
