+++ atomix/cluster/src/test/java/io/atomix/raft/roles/LeaderRoleTest.java
@@ -36,7 +36,7 @@ import io.atomix.raft.storage.log.RaftLog;
 import io.atomix.raft.storage.log.entry.ApplicationEntry;
 import io.atomix.raft.storage.log.entry.RaftEntry;
 import io.atomix.raft.storage.log.entry.RaftLogEntry;
-import io.atomix.raft.zeebe.ValidationResult;
+import io.atomix.raft.zeebe.EntryValidator.ValidationResult;
 import io.atomix.raft.zeebe.ZeebeLogAppender.AppendListener;
 import io.atomix.raft.zeebe.util.TestAppender;
 import io.atomix.utils.concurrent.SingleThreadContext;
@@ -84,7 +84,7 @@ public class LeaderRoleTest {
 
     final ReceivableSnapshotStore persistedSnapshotStore = mock(ReceivableSnapshotStore.class);
     when(context.getPersistedSnapshotStore()).thenReturn(persistedSnapshotStore);
-    when(context.getEntryValidator()).thenReturn((a, b) -> ValidationResult.success());
+    when(context.getEntryValidator()).thenReturn((a, b) -> ValidationResult.ok());
     when(context.getStorage()).thenReturn(RaftStorage.builder().withMaxSegmentSize(1024).build());
 
     leaderRole = new LeaderRole(context);
@@ -322,7 +322,7 @@ public class LeaderRoleTest {
                 latch.countDown();
                 return ValidationResult.failure("expected");
               }
-              return ValidationResult.success();
+              return ValidationResult.ok();
             });
     leaderRole = new LeaderRole(context);
